Code coverage was not working in HTTP proc in Java Issue: 201799
