+++ gxweb/src/main/java/com/genexus/webpanels/GXWebObjectBase.java
@@ -6,8 +6,6 @@ import com.genexus.*;
 import com.genexus.configuration.ConfigurationManager;
 import com.genexus.diagnostics.core.ILogger;
 import com.genexus.diagnostics.core.LogManager;
-import com.genexus.diagnostics.GXDebugInfo;
-import com.genexus.diagnostics.GXDebugManager;
 import com.genexus.internet.HttpAjaxContext;
 import com.genexus.internet.GXInternetConstants;
 import com.genexus.internet.HttpContext;
@@ -216,41 +214,6 @@ public abstract class GXWebObjectBase extends GXObjectBase implements GXInternet
      {
          return String.format("%s%s", cmpCtx, this.getPgmname().toUpperCase());
      }
-
-	private GXDebugInfo dbgInfo = null;
-	protected void trkCleanup()
-	{
-		if(dbgInfo != null)
-			dbgInfo.onCleanup();
-	}
-
-	protected void initialize(int objClass, int objId, int dbgLines, long hash)
-	{
-		dbgInfo = GXDebugManager.getInstance().getDbgInfo(context, objClass, objId, dbgLines, hash);
-	}
-
-	protected void trk(int lineNro)
-	{
-		if(dbgInfo != null)
-			dbgInfo.trk(lineNro);
-	}
-
-	protected void trk(int lineNro, int lineNro2)
-	{
-		if(dbgInfo != null)
-			dbgInfo.trk(lineNro, lineNro2);
-	}
-
-	protected void trkrng(int lineNro, int lineNro2)
-	{
-		trkrng(lineNro, 0, lineNro2, 0);
-	}
-
-	protected void trkrng(int lineNro, int colNro, int lineNro2, int colNro2)
-	{
-		if(dbgInfo != null)
-			dbgInfo.trkRng(lineNro, colNro, lineNro2, colNro2);
-	}
 
 	protected void callWebObject(String url)
 	{
